+++ src/DecagonES2.h
@@ -47,17 +47,17 @@ class DecagonES2 : public DecagonSDI12
 {
 public:
     // Constructors with overloads
-    DecagonES2(char SDI12address, int powerPin, int dataPin, int measurementsToAverage = 1)
+    DecagonES2(char SDI12address, int8_t powerPin, int8_t dataPin, uint8_t measurementsToAverage = 1)
      : DecagonSDI12(SDI12address, powerPin, dataPin, measurementsToAverage,
                     F("DecagonES2"), ES2_NUM_VARIABLES,
                     ES2_WARM_UP_TIME_MS, ES2_STABILIZATION_TIME_MS, ES2_MEASUREMENT_TIME_MS)
     {}
-    DecagonES2(char *SDI12address, int powerPin, int dataPin, int measurementsToAverage = 1)
+    DecagonES2(char *SDI12address, int8_t powerPin, int8_t dataPin, uint8_t measurementsToAverage = 1)
      : DecagonSDI12(SDI12address, powerPin, dataPin, measurementsToAverage,
                     F("DecagonES2"), ES2_NUM_VARIABLES,
                     ES2_WARM_UP_TIME_MS, ES2_STABILIZATION_TIME_MS, ES2_MEASUREMENT_TIME_MS)
     {}
-    DecagonES2(int SDI12address, int powerPin, int dataPin, int measurementsToAverage = 1)
+    DecagonES2(int SDI12address, int8_t powerPin, int8_t dataPin, uint8_t measurementsToAverage = 1)
      : DecagonSDI12(SDI12address, powerPin, dataPin, measurementsToAverage,
                     F("DecagonES2"), ES2_NUM_VARIABLES,
                     ES2_WARM_UP_TIME_MS, ES2_STABILIZATION_TIME_MS, ES2_MEASUREMENT_TIME_MS)
