+++ src/MaximDS3231.h
@@ -43,7 +43,7 @@ class MaximDS3231 : public Sensor
 {
 public:
     // Only input is the number of readings to average
-    MaximDS3231(int measurementsToAverage = 1)
+    MaximDS3231(uint8_t measurementsToAverage = 1)
     : Sensor(F("MaximDS3231"), DS3231_NUM_VARIABLES,
              DS3231_WARM_UP_TIME_MS, DS3231_STABILIZATION_TIME_MS, DS3231_MEASUREMENT_TIME_MS,
              -1, -1, measurementsToAverage)
